requirements.txt
streamlit==1.25.0
pytorch==2.13.0
numpy==1.23.5
pandas==1.5.3
opencv-python==4.8.0.74
matplotlib==3.7.1
Pillow==9.5.0
scikit-learn==1.2.2
